check locale storage type before parsing
