+++ app/[locale]/Components/hooks/useAuthCheck.tsx
@@ -3,16 +3,18 @@
 import { useEffect, useState } from "react";
 import { useRouter } from "../../../../i18n/routing";
 
-export function useAuthRedirect() {
+export function useAuthCheck() {
   const [loading, setLoading] = useState(true);
   const router = useRouter();
 
   useEffect(() => {
-    const isAuthenticated = JSON.parse(localStorage.getItem("isAuthenticated"));
-    if (isAuthenticated) {
-      router.push("/home");
+    const checkAuth = localStorage?.getItem("isAuthenticated");
+    const isAuthenticated = checkAuth && JSON.parse(checkAuth);
+    if (!isAuthenticated) {
+      router.push("/login");
     } else {
       setLoading(false);
     }
   }, [router]);
+  return loading;
 }
